Lesson3: Q1b complete - sscanf parsing + correct float average (verified with uneven test data)

+++ work/lesson3/q1b.c
@@ -22,7 +22,7 @@ int main() {
 
     }
 
-    average =  sum / count;
+    average =  (float)sum / count;
 
     printf("average:%f", average);
